Remove codes from api/stats
