+++ pages/api/stats/index.ts
@@ -1,5 +1,5 @@
 import AchievementInfo from '@root/constants/achievementInfo';
-import Direction, { getDirectionFromCode } from '@root/constants/direction';
+import Direction from '@root/constants/direction';
 import getDifficultyEstimate from '@root/helpers/getDifficultyEstimate';
 import User from '@root/models/db/user';
 import { AttemptContext } from '@root/models/schemas/playAttemptSchema';
@@ -35,8 +35,7 @@ export default withAuth({
   GET: {},
   PUT: {
     body: {
-      codes: ValidArray(false),
-      directions: ValidArray(false),
+      directions: ValidArray(),
       levelId: ValidObjectId(),
       matchId: ValidType('string', false),
     }
@@ -47,15 +46,15 @@ export default withAuth({
 
     return res.status(200).json(stats);
   } else if (req.method === 'PUT') {
-    const { codes, directions, levelId, matchId } = req.body;
+    const { directions, levelId, matchId } = req.body;
 
-    if (!directions && !codes) {
-      return res.status(400).json({
-        error: 'No directions or codes provided',
-      });
+    for (const direction of directions) {
+      if (!(direction in Direction)) {
+        return res.status(400).json({
+          error: `Invalid direction provided: ${direction}`,
+        });
+      }
     }
-
-    const dirs: Direction[] = directions ?? codes.map((code: string) => getDirectionFromCode(code));
 
     const ts = TimerUtil.getTs();
     const session = await mongoose.startSession();
@@ -78,13 +77,13 @@ export default withAuth({
           throw new Error(resTrack.json.error);
         }
 
-        if (!validateSolution(dirs, level)) {
+        if (!validateSolution(directions, level)) {
           resTrack.status = 400;
           resTrack.json.error = `Invalid solution provided for level ${levelId}`;
           throw new Error(resTrack.json.error);
         }
 
-        const moves = dirs.length;
+        const moves = directions.length;
 
         // ensure no stats are saved for draft levels
         if (level.isDraft || level.leastMoves === 0) {
